+++ sv-parser/examples/module_list.rs
@@ -14,7 +14,7 @@ fn main() {
     let includes: Vec<PathBuf> = Vec::new();
 
     // Parse
-    let result = parse_sv(&path, &defines, &includes, false);
+    let result = parse_sv(&path, &defines, &includes, false, false);
 
     if let Ok((syntax_tree, _)) = result {
         // &SyntexTree is iterable