+++ sv-parser/src/lib.rs
@@ -97,17 +97,23 @@ pub fn parse_sv<T: AsRef<Path>, U: AsRef<Path>, V: BuildHasher>(
     pre_defines: &HashMap<String, Option<Define>, V>,
     include_paths: &[U],
     ignore_include: bool,
+    allow_incomplete: bool,
 ) -> Result<(SyntaxTree, Defines), Error> {
     let (text, defines) = preprocess(path, pre_defines, include_paths, false, ignore_include)?;
-    parse_sv_pp(text, defines)
+    parse_sv_pp(text, defines, allow_incomplete)
 }
 
 pub fn parse_sv_pp(
     text: PreprocessedText,
     defines: Defines,
+    allow_incomplete: bool,
 ) -> Result<(SyntaxTree, Defines), Error> {
     let span = Span::new_extra(text.text(), SpanInfo::default());
-    let result = all_consuming(sv_parser)(span);
+    let result = if allow_incomplete {
+        sv_parser(span)
+    } else {
+        all_consuming(sv_parser)(span)
+    };
     match result {
         Ok((_, x)) => Ok((
             SyntaxTree {
@@ -142,6 +148,7 @@ pub fn parse_sv_str<T: AsRef<Path>, U: AsRef<Path>, V: BuildHasher>(
     pre_defines: &HashMap<String, Option<Define>, V>,
     include_paths: &[U],
     ignore_include: bool,
+    allow_incomplete: bool,
 ) -> Result<(SyntaxTree, Defines), Error> {
     let (text, defines) = preprocess_str(
         s,
@@ -152,7 +159,7 @@ pub fn parse_sv_str<T: AsRef<Path>, U: AsRef<Path>, V: BuildHasher>(
         false,
         0,
     )?;
-    parse_sv_pp(text, defines)
+    parse_sv_pp(text, defines, allow_incomplete)
 }
 
 pub fn parse_lib<T: AsRef<Path>, U: AsRef<Path>, V: BuildHasher>(
@@ -160,9 +167,10 @@ pub fn parse_lib<T: AsRef<Path>, U: AsRef<Path>, V: BuildHasher>(
     pre_defines: &HashMap<String, Option<Define>, V>,
     include_paths: &[U],
     ignore_include: bool,
+    allow_incomplete: bool,
 ) -> Result<(SyntaxTree, Defines), Error> {
     let (text, defines) = preprocess(path, pre_defines, include_paths, false, ignore_include)?;
-    parse_lib_pp(text, defines)
+    parse_lib_pp(text, defines, allow_incomplete)
 }
 
 pub fn parse_lib_str<T: AsRef<Path>, U: AsRef<Path>, V: BuildHasher>(
@@ -171,6 +179,7 @@ pub fn parse_lib_str<T: AsRef<Path>, U: AsRef<Path>, V: BuildHasher>(
     pre_defines: &HashMap<String, Option<Define>, V>,
     include_paths: &[U],
     ignore_include: bool,
+    allow_incomplete: bool,
 ) -> Result<(SyntaxTree, Defines), Error> {
     let (text, defines) = preprocess_str(
         s,
@@ -181,15 +190,20 @@ pub fn parse_lib_str<T: AsRef<Path>, U: AsRef<Path>, V: BuildHasher>(
         false,
         0,
     )?;
-    parse_lib_pp(text, defines)
+    parse_lib_pp(text, defines, allow_incomplete)
 }
 
 pub fn parse_lib_pp(
     text: PreprocessedText,
     defines: Defines,
+    allow_incomplete: bool,
 ) -> Result<(SyntaxTree, Defines), Error> {
     let span = Span::new_extra(text.text(), SpanInfo::default());
-    let result = all_consuming(lib_parser)(span);
+    let result = if allow_incomplete {
+        all_consuming(lib_parser)(span)
+    } else {
+        lib_parser(span)
+    };
     match result {
         Ok((_, x)) => Ok((
             SyntaxTree {
@@ -258,7 +272,7 @@ mod test {
     fn test() {
         let src = "/* comment */";
         let (syntax_tree, _) =
-            parse_sv_str(src, PathBuf::from(""), &HashMap::new(), &[""], false).unwrap();
+            parse_sv_str(src, PathBuf::from(""), &HashMap::new(), &[""], false, false).unwrap();
         let comment = unwrap_node!(&syntax_tree, Comment);
         assert!(comment.is_some());
     }